Fix a memory leak when the sequence buffer wraps
If the sequence buffers wrap around while a fragmented packet is still
waiting for reassembly and we receive a non-fragmented packet that cause
us to advance the fragmented sequence buffer past the incomplete
fragmented packet, we would not call a cleanup function on the partially
reassembled packet, leaving the packet_data to leak.

This changes the call to `reliable_sequence_buffer_advance` on the
reassembly sequence buffer in the non-fragmented code path to a newly
introduced `reliable_sequence_buffer_advance_with_cleanup` which does
the correct cleanup on the reassembly sequence buffer entries.