Switch header reading code to log with DEBUG
This switches log messages in the receive code path to use
`RELIABLE_LOG_LEVEL_DEBUG` instead of `RELIABLE_LOG_LEVEL_ERROR` until
after we've determined that the header seems sensible. This should make
it so that it's unlikely for a random datagram across the internet
causes error logging in your application.